+++ R/NCObject.R
@@ -56,11 +56,11 @@ NCObject <- R6::R6Class("NCObject",
     #'
     #' This method returns netCDF object attributes.
     #'
-    #' @param att Vector of character strings of attributes to return.
+    #' @param att Vector of attribute names whose values to return.
     #' @param field The field of the `data.frame` to return values from. This
     #' must be "value" (default), "type" or "length".
     #' @returns A vector of values from the `data.frame`, named with the `att`
-    #' value.
+    #' value, or `character(0)` if the attribute is not found.
     attribute = function(att, field = "value") {
       atts <- self$attributes
       if (!nrow(atts)) return(character(0L))